Refactor package of Tag
Refactored package of Tag
