+++ src/main/java/seedu/address/logic/Messages.java
@@ -1,6 +1,6 @@
 package seedu.address.logic;
 
-import static seedu.address.model.tag.Tag.MAXIMUM_TAGS_PER_PERSON;
+import static seedu.address.model.person.Tag.MAXIMUM_TAGS_PER_PERSON;
 
 import java.util.Set;
 import java.util.stream.Collectors;
