+++ src/main/java/seedu/address/logic/commands/TagCommand.java
@@ -7,7 +7,7 @@ import static seedu.address.logic.parser.CliSyntax.PREFIX_ADD_TAG;
 import static seedu.address.logic.parser.CliSyntax.PREFIX_DELETE_TAG;
 import static seedu.address.model.Model.PREDICATE_SHOW_ALL_PERSONS;
 import static seedu.address.model.person.Person.createPersonWithUpdatedTags;
-import static seedu.address.model.tag.Tag.MAXIMUM_TAGS_PER_PERSON;
+import static seedu.address.model.person.Tag.MAXIMUM_TAGS_PER_PERSON;
 
 import java.util.HashSet;
 import java.util.List;
@@ -20,7 +20,7 @@ import seedu.address.logic.Messages;
 import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.model.Model;
 import seedu.address.model.person.Person;
-import seedu.address.model.tag.Tag;
+import seedu.address.model.person.Tag;
 
 /**
  * Adds or deletes tags of a person identified using it's displayed index from the address book.
