+++ src/main/java/seedu/address/logic/parser/ParserUtil.java
@@ -16,8 +16,8 @@ import seedu.address.model.person.EmptyAddress;
 import seedu.address.model.person.Name;
 import seedu.address.model.person.NonEmptyAddress;
 import seedu.address.model.person.Phone;
+import seedu.address.model.person.Tag;
 import seedu.address.model.priority.Priority;
-import seedu.address.model.tag.Tag;
 
 /**
  * Contains utility methods used for parsing strings in the various *Parser classes.
