+++ src/main/java/seedu/address/model/person/Person.java
@@ -10,7 +10,6 @@ import java.util.Set;
 
 import seedu.address.commons.util.ToStringBuilder;
 import seedu.address.model.priority.Priority;
-import seedu.address.model.tag.Tag;
 
 /**
  * Represents a Person in the address book.
