+++ src/test/java/seedu/address/logic/commands/TagCommandTest.java
@@ -25,7 +25,7 @@ import seedu.address.model.Model;
 import seedu.address.model.ModelManager;
 import seedu.address.model.UserPrefs;
 import seedu.address.model.person.Person;
-import seedu.address.model.tag.Tag;
+import seedu.address.model.person.Tag;
 import seedu.address.testutil.PersonBuilder;
 import seedu.address.testutil.UpdatePersonTagsDescriptorBuilder;
