+++ src/test/java/seedu/address/logic/parser/TagCommandParserTest.java
@@ -20,7 +20,7 @@ import seedu.address.commons.core.index.Index;
 import seedu.address.logic.commands.Command;
 import seedu.address.logic.commands.TagCommand;
 import seedu.address.logic.commands.TagCommand.UpdatePersonTagsDescriptor;
-import seedu.address.model.tag.Tag;
+import seedu.address.model.person.Tag;
 import seedu.address.testutil.UpdatePersonTagsDescriptorBuilder;
 
 public class TagCommandParserTest {
