+++ src/test/java/seedu/address/model/priority/PriorityTest.java
@@ -10,7 +10,7 @@ import static seedu.address.testutil.Assert.assertThrows;
 
 import org.junit.jupiter.api.Test;
 
-import seedu.address.model.tag.Tag;
+import seedu.address.model.person.Tag;
 
 public class PriorityTest {
     @Test
